+++ api-internal/src/main/java/at/pavlov/internal/enums/BreakCause.java
@@ -1,4 +1,4 @@
-package at.pavlov.cannons.Enum;
+package at.pavlov.internal.enums;
 
 
 public enum BreakCause {
@@ -12,8 +12,6 @@ public enum BreakCause {
     Dismantling ("dismantling"),
     Other ("other");
 
-
-
     private final String str;
 
     BreakCause(String str)
@@ -25,5 +23,4 @@ public enum BreakCause {
     {
         return str;
     }
-
 }
