+++ api-internal/src/main/java/at/pavlov/internal/enums/DamageType.java
@@ -1,4 +1,4 @@
-package at.pavlov.cannons.Enum;
+package at.pavlov.internal.enums;
 
 public enum DamageType {
     EXPLOSION,
