+++ api-internal/src/main/java/at/pavlov/internal/enums/FakeBlockType.java
@@ -1,8 +1,11 @@
 package at.pavlov.internal.enums;
 
+import lombok.Getter;
+
 /**
  * Created by Peter on 25.04.2014.
  */
+@Getter
 public enum FakeBlockType {
     AIMING("Aiming"),
     EXPLOSION("Explosion"),
@@ -17,8 +20,4 @@ public enum FakeBlockType {
         this.str = type;
     }
 
-
-    public String getStr() {
-        return str;
-    }
 }
