+++ api-internal/src/main/java/at/pavlov/internal/enums/SelectCannon.java
@@ -1,4 +1,4 @@
-package at.pavlov.cannons.Enum;
+package at.pavlov.internal.enums;
 
 public enum SelectCannon {
     OBSERVER,
