+++ cannons-bukkit/src/main/java/at/pavlov/cannons/Cannons.java
@@ -1,7 +1,7 @@
 package at.pavlov.cannons;
 
 import at.pavlov.cannons.API.CannonsAPI;
-import at.pavlov.cannons.Enum.MessageEnum;
+import at.pavlov.internal.enums.MessageEnum;
 import at.pavlov.cannons.cannon.Cannon;
 import at.pavlov.cannons.cannon.CannonDesign;
 import at.pavlov.cannons.cannon.CannonManager;
