+++ cannons-bukkit/src/main/java/at/pavlov/cannons/config/UserMessages.java
@@ -1,7 +1,7 @@
 package at.pavlov.cannons.config;
 
 import at.pavlov.cannons.Cannons;
-import at.pavlov.cannons.Enum.MessageEnum;
+import at.pavlov.internal.enums.MessageEnum;
 import at.pavlov.cannons.cannon.Cannon;
 import at.pavlov.cannons.cannon.CannonManager;
 import at.pavlov.cannons.projectile.Projectile;
