+++ cannons-bukkit/src/main/java/at/pavlov/cannons/container/SpawnEntityHolder.java
@@ -1,6 +1,6 @@
 package at.pavlov.cannons.container;
 
-import at.pavlov.cannons.Enum.EntityDataType;
+import at.pavlov.internal.enums.EntityDataType;
 import at.pavlov.internal.MaxMinRandom;
 import com.cryptomorin.xseries.XEntityType;
 import com.google.gson.JsonElement;
