+++ cannons-bukkit/src/main/java/at/pavlov/cannons/dao/wrappers/FireTaskWrapper.java
@@ -1,9 +1,9 @@
 package at.pavlov.cannons.dao.wrappers;
 
 import at.pavlov.cannons.Cannons;
-import at.pavlov.cannons.Enum.BreakCause;
+import at.pavlov.internal.enums.BreakCause;
 import at.pavlov.internal.enums.FakeBlockType;
-import at.pavlov.cannons.Enum.ProjectileCause;
+import at.pavlov.internal.enums.ProjectileCause;
 import at.pavlov.cannons.cannon.Cannon;
 import at.pavlov.cannons.cannon.CannonDesign;
 import at.pavlov.cannons.cannon.CannonManager;
