+++ cannons-bukkit/src/main/java/at/pavlov/cannons/dao/wrappers/RemoveTaskWrapper.java
@@ -1,6 +1,6 @@
 package at.pavlov.cannons.dao.wrappers;
 
-import at.pavlov.cannons.Enum.BreakCause;
+import at.pavlov.internal.enums.BreakCause;
 import at.pavlov.cannons.cannon.Cannon;
 
 public record RemoveTaskWrapper(
