+++ cannons-bukkit/src/main/java/at/pavlov/cannons/event/CannonUseEvent.java
@@ -1,6 +1,6 @@
 package at.pavlov.cannons.event;
 
-import at.pavlov.cannons.Enum.InteractAction;
+import at.pavlov.internal.enums.InteractAction;
 import at.pavlov.cannons.cannon.Cannon;
 import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
