+++ cannons-bukkit/src/main/java/at/pavlov/cannons/interfaces/holders/ObserverMapHolder.java
@@ -1,6 +1,6 @@
 package at.pavlov.cannons.interfaces.holders;
 
-import at.pavlov.cannons.Enum.MessageEnum;
+import at.pavlov.internal.enums.MessageEnum;
 import org.apache.commons.lang3.Validate;
 import org.bukkit.entity.Player;
