+++ cannons-bukkit/src/main/java/at/pavlov/cannons/projectile/FlyingProjectile.java
@@ -1,6 +1,6 @@
 package at.pavlov.cannons.projectile;
 
-import at.pavlov.cannons.Enum.ProjectileCause;
+import at.pavlov.internal.enums.ProjectileCause;
 import at.pavlov.cannons.container.MovingObject;
 import io.papermc.lib.PaperLib;
 import org.bukkit.Bukkit;
