+++ cannons-bukkit/src/main/java/at/pavlov/cannons/projectile/ProjectileManager.java
@@ -2,7 +2,7 @@ package at.pavlov.cannons.projectile;
 
 import at.pavlov.cannons.Cannons;
 import at.pavlov.cannons.CreateExplosion;
-import at.pavlov.cannons.Enum.ProjectileCause;
+import at.pavlov.internal.enums.ProjectileCause;
 import at.pavlov.cannons.dao.AsyncTaskManager;
 import org.apache.commons.lang3.Validate;
 import org.bukkit.Location;
