+++ cannons-bukkit/src/main/java/at/pavlov/cannons/utils/CannonSelector.java
@@ -1,12 +1,11 @@
 package at.pavlov.cannons.utils;
 
 import at.pavlov.cannons.Cannons;
-import at.pavlov.cannons.Enum.MessageEnum;
-import at.pavlov.cannons.Enum.SelectCannon;
+import at.pavlov.internal.enums.MessageEnum;
+import at.pavlov.internal.enums.SelectCannon;
 import at.pavlov.cannons.cannon.Cannon;
 import at.pavlov.cannons.cannon.CannonManager;
 import at.pavlov.cannons.config.UserMessages;
-import at.pavlov.cannons.exchange.EmptyExchanger;
 import lombok.Getter;
 import org.bukkit.OfflinePlayer;
 import org.bukkit.block.Block;
